rename `GlideGrid` to `Grid` (#1462)
