+++ packages/hash/frontend/src/components/grid/utils/draw-cell-fade-out-gradient.ts
