+++ packages/hash/frontend/src/components/grid/utils/draw-chip-with-icon.ts
