+++ packages/hash/frontend/src/components/grid/utils/draw-round-rect.ts
