+++ packages/hash/frontend/src/components/grid/utils/draw-text-with-icon.ts
