+++ packages/hash/frontend/src/components/grid/utils/interactable-manager/types.ts
