+++ packages/hash/frontend/src/components/grid/utils/interactable-manager/utils.ts
