+++ packages/hash/frontend/src/components/grid/utils/use-draw-cell.ts
