+++ packages/hash/frontend/src/components/grid/utils/use-grid-tooltip/draw-interactable-tooltip-icons.ts
