+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-id].page/entity-editor/properties-section/property-table/cells/summary-chip-cell.tsx
@@ -3,9 +3,9 @@ import {
   CustomRenderer,
   GridCellKind,
 } from "@glideapps/glide-data-grid";
-import { getYCenter } from "../../../../../../../../components/GlideGlid/utils";
-import { drawCellFadeOutGradient } from "../../../../../../../../components/GlideGlid/utils/draw-cell-fade-out-gradient";
-import { drawRoundRect } from "../../../../../../../../components/GlideGlid/utils/draw-round-rect";
+import { getYCenter } from "../../../../../../../../components/grid/utils";
+import { drawCellFadeOutGradient } from "../../../../../../../../components/grid/utils/draw-cell-fade-out-gradient";
+import { drawRoundRect } from "../../../../../../../../components/grid/utils/draw-round-rect";
 
 export interface SummaryChipCellProps {
   readonly kind: "summary-chip-cell";
