+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-id].page/entity-editor/properties-section/property-table/cells/value-cell/types.ts
@@ -1,5 +1,5 @@
 import { CustomCell, ProvideEditorComponent } from "@glideapps/glide-data-grid";
-import { TooltipCellProps } from "../../../../../../../../../components/GlideGlid/utils/use-grid-tooltip/types";
+import { TooltipCellProps } from "../../../../../../../../../components/grid/utils/use-grid-tooltip/types";
 import { PropertyRow } from "../../types";
 
 export interface ValueCellProps extends TooltipCellProps {
