+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-id].page/entity-editor/properties-section/property-table/fill-row-data-indent-calculations.ts
@@ -1,4 +1,4 @@
-import { VerticalIndentationLineDir } from "../../../../../../../components/GlideGlid/utils/draw-vertical-indentation-line";
+import { VerticalIndentationLineDir } from "../../../../../../../components/grid/utils/draw-vertical-indentation-line";
 import { PropertyRow } from "./types";
 
 const markUpperHalvesOfIndentationLines = (rowData: PropertyRow[]) => {
