+++ packages/hash/frontend/src/pages/[account-slug]/entities/[entity-id].page/entity-editor/properties-section/property-table/get-tooltips-of-property-row.ts
@@ -1,4 +1,4 @@
-import { CellTooltipData } from "../../../../../../../components/GlideGlid/utils/use-grid-tooltip/types";
+import { CellTooltipData } from "../../../../../../../components/grid/utils/use-grid-tooltip/types";
 import { PropertyRow } from "./types";
 
 export const getTooltipsOfPropertyRow = (
