+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/tabs/entities-tab.tsx
@@ -8,10 +8,9 @@ import {
 import { Box, Paper, Stack } from "@mui/material";
 import { FunctionComponent, useCallback, useMemo, useState } from "react";
 import { faCircleQuestion } from "@fortawesome/free-regular-svg-icons";
-import { GlideGrid } from "../../../../../components/GlideGlid/glide-grid";
 import { SectionWrapper } from "../../../shared/section-wrapper";
 import { WhiteChip } from "../../../shared/white-chip";
-import { blankCell } from "../../../../../components/GlideGlid/utils";
+import { blankCell } from "../../../../../components/grid/utils";
 import { HomeIcon } from "../../../../../shared/icons/home-icon";
 import { EarthIcon } from "../../../../../shared/icons/earth-icon";
 import { renderTextIconCell } from "../text-icon-cell";
@@ -19,6 +18,7 @@ import { useRouteNamespace } from "../use-route-namespace";
 import { TypeEntitiesRow, useEntitiesTable } from "../use-entities-table";
 import { useEntityTypeEntities } from "../use-entity-type-entities";
 import { useEntityType } from "../use-entity-type";
+import { Grid } from "../../../../../components/grid/grid";
 
 export const EntitiesTab: FunctionComponent = () => {
   const entityType = useEntityType();
@@ -126,7 +126,7 @@ export const EntitiesTab: FunctionComponent = () => {
         }
       >
         <Paper sx={{ overflow: "hidden" }}>
-          <GlideGrid
+          <Grid
             showSearch={showSearch}
             onSearchClose={() => setShowSearch(false)}
             columns={columns}
