+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/text-icon-cell.tsx
@@ -4,8 +4,8 @@ import {
   GridCellKind,
 } from "@glideapps/glide-data-grid";
 import type { CustomIcon } from "@glideapps/glide-data-grid/dist/ts/data-grid/data-grid-sprites";
-import { getCellHorizontalPadding } from "../../../../components/GlideGlid/utils";
-import { drawTextWithIcon } from "../../../../components/GlideGlid/utils/draw-text-with-icon";
+import { getCellHorizontalPadding } from "../../../../components/grid/utils";
+import { drawTextWithIcon } from "../../../../components/grid/utils/draw-text-with-icon";
 
 export interface TextIconCellProps {
   readonly kind: "text-icon-cell";
